(fix) copy md2.h to correct place
Previously, we had a patch for various openssl source files where
md2.h was included to point it to where md2.h is present originally.
however, after noticing that all other header files are under openssl
and not in their original path, it is better to copy the md2.h to the
correct location.