(FACT-914) Update Boost to 1.57
Update Boost to 1.57 to fix boost::filesystem's dependence on
std::locale. On Solaris with GCC, std::locale is broken and this caused
an exception to be thrown unless facter is run with LC_ALL=C.

Required a minor tweak to compilation settings to build Boost 1.57 with
GCC on Solaris - Boost.Log defines _XOPEN_SOURCE=500, which is
incompatible with GCC's headers when building with c99 compatibility. We
redefine to _XOPEN_SOURCE=600, which works on Solaris 10 and 11.

Use https for Git repos, because wth weren't we?